Add files via upload
The following changes were made:
1. Comments for constants and readonly variables now use the verb "contain" unless the latter refers to a lamba procedure.
2. Comments for enumerations use the verb "list" and comments for variables use the verb "contain".
3. All references to the terms "chunk" and "data chunk" regarding the Cartooners executable file have been changed to use the term "region".

Note:
While Cartooners File Viewer can provide a lot of information about Cartooners' files it is far from complete or perfect. I have started a project attempting to disassamble the executable code but should anything ever come of it, it may take months. This is not a high-priority project for me.  Feed-back is, however, always welcome.